Moving Customization Testing to Correct Tab -All customization aspects have been moved to the correct tab -Customization tab top has had its name changed. -Renamed files that have become customization tab -Added new buttons that change colors
