+++ app/src/main/java/com/example/testing/ui/home/HomeFragment.kt
@@ -1,11 +1,9 @@
 package com.example.testing.ui.home
 
-import android.graphics.Color
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.example.testing.databinding.FragmentHomeBinding
@@ -28,21 +26,10 @@ class HomeFragment: Fragment() {
             ViewModelProvider(this).get(HomeViewModel::class.java)
 
         _binding = FragmentHomeBinding.inflate(inflater, container, false)
-        val root: View = binding.root
 
-        val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner) {
-            textView.text = it
-        }
-        return root
 
-    }
+        return binding.root
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.colorButton.setOnClickListener{
-            binding.enemy.setColorFilter(Color.BLACK)
-        }
     }
 
     override fun onDestroyView() {
